+++ CHANGELOG.md
@@ -8,6 +8,7 @@ we hit release version 1.0.0.
 ## [UNRELEASED] - YYYY-MM-DD
 
 ### Added
+- enabled parsing geometry.in files from FHIaims
 - added `batched_indices` for memory-reduced location of array values
 - enabled manifold extractions `sisl.physics.yield_manifolds`
 - enabled center of mass for periodic systems (chooses *best* COM)